+++ layers/vk_layer_utils.h
@@ -22,6 +22,7 @@
 #pragma once
 #include <stdbool.h>
 #include <vector>
+#include "vk_format_utils.h"
 #include "vk_layer_logging.h"
 
 #ifndef WIN32
@@ -35,54 +36,6 @@ extern "C" {
 #endif
 
 #define VK_LAYER_API_VERSION VK_MAKE_VERSION(1, 0, VK_HEADER_VERSION)
-typedef enum VkFormatCompatibilityClass {
-    VK_FORMAT_COMPATIBILITY_CLASS_NONE_BIT = 0,
-    VK_FORMAT_COMPATIBILITY_CLASS_8_BIT = 1,
-    VK_FORMAT_COMPATIBILITY_CLASS_16_BIT = 2,
-    VK_FORMAT_COMPATIBILITY_CLASS_24_BIT = 3,
-    VK_FORMAT_COMPATIBILITY_CLASS_32_BIT = 4,
-    VK_FORMAT_COMPATIBILITY_CLASS_48_BIT = 5,
-    VK_FORMAT_COMPATIBILITY_CLASS_64_BIT = 6,
-    VK_FORMAT_COMPATIBILITY_CLASS_96_BIT = 7,
-    VK_FORMAT_COMPATIBILITY_CLASS_128_BIT = 8,
-    VK_FORMAT_COMPATIBILITY_CLASS_192_BIT = 9,
-    VK_FORMAT_COMPATIBILITY_CLASS_256_BIT = 10,
-    VK_FORMAT_COMPATIBILITY_CLASS_BC1_RGB_BIT = 11,
-    VK_FORMAT_COMPATIBILITY_CLASS_BC1_RGBA_BIT = 12,
-    VK_FORMAT_COMPATIBILITY_CLASS_BC2_BIT = 13,
-    VK_FORMAT_COMPATIBILITY_CLASS_BC3_BIT = 14,
-    VK_FORMAT_COMPATIBILITY_CLASS_BC4_BIT = 15,
-    VK_FORMAT_COMPATIBILITY_CLASS_BC5_BIT = 16,
-    VK_FORMAT_COMPATIBILITY_CLASS_BC6H_BIT = 17,
-    VK_FORMAT_COMPATIBILITY_CLASS_BC7_BIT = 18,
-    VK_FORMAT_COMPATIBILITY_CLASS_ETC2_RGB_BIT = 19,
-    VK_FORMAT_COMPATIBILITY_CLASS_ETC2_RGBA_BIT = 20,
-    VK_FORMAT_COMPATIBILITY_CLASS_ETC2_EAC_RGBA_BIT = 21,
-    VK_FORMAT_COMPATIBILITY_CLASS_EAC_R_BIT = 22,
-    VK_FORMAT_COMPATIBILITY_CLASS_EAC_RG_BIT = 23,
-    VK_FORMAT_COMPATIBILITY_CLASS_ASTC_4X4_BIT = 24,
-    VK_FORMAT_COMPATIBILITY_CLASS_ASTC_5X4_BIT = 25,
-    VK_FORMAT_COMPATIBILITY_CLASS_ASTC_5X5_BIT = 26,
-    VK_FORMAT_COMPATIBILITY_CLASS_ASTC_6X5_BIT = 27,
-    VK_FORMAT_COMPATIBILITY_CLASS_ASTC_6X6_BIT = 28,
-    VK_FORMAT_COMPATIBILITY_CLASS_ASTC_8X5_BIT = 29,
-    VK_FORMAT_COMPATIBILITY_CLASS_ASTC_8X6_BIT = 20,
-    VK_FORMAT_COMPATIBILITY_CLASS_ASTC_8X8_BIT = 31,
-    VK_FORMAT_COMPATIBILITY_CLASS_ASTC_10X5_BIT = 32,
-    VK_FORMAT_COMPATIBILITY_CLASS_ASTC_10X6_BIT = 33,
-    VK_FORMAT_COMPATIBILITY_CLASS_ASTC_10X8_BIT = 34,
-    VK_FORMAT_COMPATIBILITY_CLASS_ASTC_10X10_BIT = 35,
-    VK_FORMAT_COMPATIBILITY_CLASS_ASTC_12X10_BIT = 36,
-    VK_FORMAT_COMPATIBILITY_CLASS_ASTC_12X12_BIT = 37,
-    VK_FORMAT_COMPATIBILITY_CLASS_D16_BIT = 38,
-    VK_FORMAT_COMPATIBILITY_CLASS_D24_BIT = 39,
-    VK_FORMAT_COMPATIBILITY_CLASS_D32_BIT = 30,
-    VK_FORMAT_COMPATIBILITY_CLASS_S8_BIT = 41,
-    VK_FORMAT_COMPATIBILITY_CLASS_D16S8_BIT = 42,
-    VK_FORMAT_COMPATIBILITY_CLASS_D24S8_BIT = 43,
-    VK_FORMAT_COMPATIBILITY_CLASS_D32S8_BIT = 44,
-    VK_FORMAT_COMPATIBILITY_CLASS_MAX_ENUM = 45
-} VkFormatCompatibilityClass;
 
 typedef enum VkStringErrorFlagBits {
     VK_STRING_ERROR_NONE = 0x00000000,
@@ -94,44 +47,6 @@ typedef VkFlags VkStringErrorFlags;
 VK_LAYER_EXPORT void layer_debug_actions(debug_report_data *report_data, std::vector<VkDebugReportCallbackEXT> &logging_callback,
                                          const VkAllocationCallbacks *pAllocator, const char *layer_identifier);
 
-static inline bool vk_format_is_undef(VkFormat format) { return (format == VK_FORMAT_UNDEFINED); }
-
-bool vk_format_is_depth_or_stencil(VkFormat format);
-bool vk_format_is_depth_and_stencil(VkFormat format);
-bool vk_format_is_depth_only(VkFormat format);
-bool vk_format_is_stencil_only(VkFormat format);
-
-static inline bool vk_format_is_color(VkFormat format) {
-    return !(vk_format_is_undef(format) || vk_format_is_depth_or_stencil(format));
-}
-
-static inline bool vk_format_has_depth(VkFormat format) {
-    return (vk_format_is_depth_only(format) || vk_format_is_depth_and_stencil(format));
-}
-
-static inline bool vk_format_has_stencil(VkFormat format) {
-    return (vk_format_is_stencil_only(format) || vk_format_is_depth_and_stencil(format));
-}
-
-VK_LAYER_EXPORT bool vk_format_is_compressed_ETC2_EAC(VkFormat format);
-VK_LAYER_EXPORT bool vk_format_is_compressed_ASTC_LDR(VkFormat format);
-VK_LAYER_EXPORT bool vk_format_is_compressed_BC(VkFormat format);
-VK_LAYER_EXPORT bool vk_format_is_norm(VkFormat format);
-VK_LAYER_EXPORT bool vk_format_is_unorm(VkFormat format);
-VK_LAYER_EXPORT bool vk_format_is_snorm(VkFormat format);
-VK_LAYER_EXPORT bool vk_format_is_int(VkFormat format);
-VK_LAYER_EXPORT bool vk_format_is_sint(VkFormat format);
-VK_LAYER_EXPORT bool vk_format_is_uint(VkFormat format);
-VK_LAYER_EXPORT bool vk_format_is_float(VkFormat format);
-VK_LAYER_EXPORT bool vk_format_is_srgb(VkFormat format);
-VK_LAYER_EXPORT bool vk_format_is_uscaled(VkFormat format);
-VK_LAYER_EXPORT bool vk_format_is_sscaled(VkFormat format);
-VK_LAYER_EXPORT bool vk_format_is_compressed(VkFormat format);
-VK_LAYER_EXPORT VkExtent3D vk_format_compressed_texel_block_extents(VkFormat format);
-VK_LAYER_EXPORT size_t vk_format_get_size(VkFormat format);
-VK_LAYER_EXPORT unsigned int vk_format_get_channel_count(VkFormat format);
-VK_LAYER_EXPORT VkFormatCompatibilityClass vk_format_get_compatibility_class(VkFormat format);
-VK_LAYER_EXPORT VkDeviceSize vk_safe_modulo(VkDeviceSize dividend, VkDeviceSize divisor);
 VK_LAYER_EXPORT VkStringErrorFlags vk_string_validate(const int max_length, const char *char_array);
 VK_LAYER_EXPORT bool white_list(const char *item, const char *whitelist);
