+++ tests/vkrenderframework.h
@@ -445,5 +445,7 @@ class VkPipelineObj : public vk_testing::Pipeline {
     vector<VkPipelineColorBlendAttachmentState> m_colorAttachments;
     int m_vertexBufferCount;
 };
-VkFormat find_depth_stencil_format(VkDeviceObj *device);
+
+VkFormat FindDepthStencilFormat(VkPhysicalDevice);
+
 #endif  // VKRENDERFRAMEWORK_H
